User Put methods added

+++ demo/src/main/java/com/deneme/demo/services/UserService.java
@@ -10,6 +10,8 @@ import com.deneme.demo.entities.User;
 import com.deneme.demo.repos.UserRepository;
 import com.deneme.demo.requests.LoginRequest;
 import com.deneme.demo.requests.SoftDeleteRequest;
+import com.deneme.demo.requests.UpdatePasswordRequest;
+import com.deneme.demo.requests.UpdateUserRequest;
 
 @Service
 public class UserService {
@@ -22,11 +24,15 @@ public class UserService {
 	}
 	
 	public User saveOneUser(User newUser) {
-		User user = userRepository.findByTcNo(newUser.getTcNo());
-		if(user == null)
+		if(newUser.getTcNo() != null)
 		{
-			newUser.setHashPassword(PasswordHashing.hashPassword(newUser.getHashPassword()));
-			return userRepository.save(newUser);
+			if(!userRepository.existsByTcNo(newUser.getTcNo()))
+			{
+				newUser.setHashPassword(PasswordHashing.hashPassword(newUser.getHashPassword()));
+				return userRepository.save(newUser);
+			}
+			else
+				return null;
 		}
 		else
 			return null;
@@ -37,22 +43,28 @@ public class UserService {
 	}
 	
 	public void deleteUser(Long userId) {
+		if(userId != null && userRepository.findById(userId) != null)
 		userRepository.deleteById(userId);
 	}
 	
 	
 	public User softDeleteUser(Long userId,SoftDeleteRequest deleteRequest) {
-		Optional<User> user = userRepository.findById(userId);
-		if(user.isPresent())
+		Optional<User> user;
+		if(userId != null)
 		{
-			user.get().setDeleted(deleteRequest.isDeleted());
-			user.get().setUpdatedDate(deleteRequest.getUpdatedDate());
-			User updatedUser = userRepository.save(user.get());
-			return updatedUser;
+			user = userRepository.findById(userId);
+			if(user.isPresent())
+			{
+				user.get().setDeleted(deleteRequest.isDeleted());
+				user.get().setUpdatedDate(deleteRequest.getUpdatedDate());
+				User updatedUser = userRepository.save(user.get());
+				return updatedUser;
+			}
+			else 
+				return null;
 		}
-		else {
+		else 
 			return null;
-		}
 	}
 	
 	public User loginUser(LoginRequest loginRequest) {
@@ -64,4 +76,51 @@ public class UserService {
 		}
 		return null;
 	}
+
+	public User updateUser(Long userId, UpdateUserRequest updateUser) {
+		Optional<User> user;
+		if(userId != null)
+		{
+			user = userRepository.findById(userId);
+			if(user.isPresent())
+			{
+				user.get().setName(updateUser.getName());
+				user.get().setSurname(updateUser.getSurname());
+				user.get().setBirthPlace(updateUser.getBirthPlace());
+				user.get().setFatherName(updateUser.getFatherName());
+				user.get().setMotherName(updateUser.getMotherName());
+				user.get().setPhone(updateUser.getPhone());
+				user.get().setUpdatedDate(updateUser.getUpdatedDate());
+				return userRepository.save(user.get());
+			}
+			else
+				return null;
+		}
+		else
+			return null;
+	}
+	
+	public User updatePassword(Long userId, UpdatePasswordRequest updatePasswordRequest) {
+		Optional<User> user;
+		if(userId != null)
+		{
+			user = userRepository.findById(userId);
+			if(user.isPresent())
+			{
+				if(PasswordHashing.hashPassword(updatePasswordRequest.getOldPassword()).equals(user.get().getHashPassword()))
+				{
+					user.get().setHashPassword(PasswordHashing.hashPassword(updatePasswordRequest.getNewPassword()));
+					user.get().setUpdatedDate(updatePasswordRequest.getUpdatedDate());
+					return userRepository.save(user.get());
+				}else
+					//Şifre Eşleşmiyor
+					return null;
+			}else
+				//Kullanıcı Bulunamadı
+				return null;
+		}else
+			//UserID boş
+			return null;
+	}
+			
 }
